+++ src/components/library/LibraryLanding.tsx
@@ -4,6 +4,7 @@ import LibrarySection from "./LibrarySection";
 import Break from "../general/Break";
 import LibraryNavbar from "./LibraryNavbar";
 import { Book } from "../general/helpers/generalTypes";
+import LibraryRow from "./LibraryRow";
 
 export default function LibraryLanding() {
     const [editMode, setEditMode] = useState<boolean>(false);
@@ -20,24 +21,10 @@ export default function LibraryLanding() {
     return (
         <div className="d-flex flex-wrap justify-content-start justify-content-md-center mt-5 w-100">
             {username ? <LibraryNavbar username={username} /> : null}
-            <LibrarySection
-                message="Lendo"
-                bookCategory={"reading"}
-                booksInfo={user["reading"]}
-                setProgress={setProgress}
-            />
-            <Break />
-            <LibrarySection
-                message="Planejando ler"
+            <LibraryRow
+                message="Para ler"
                 bookCategory={"to-read"}
                 booksInfo={user["to-read"]}
-                setProgress={setProgress}
-            />
-            <Break />
-            <LibrarySection
-                message="Backlog"
-                bookCategory={"backlog"}
-                booksInfo={user["backlog"]}
                 setProgress={setProgress}
             />
             <Break />