+++ src/components/library/LibraryNavbar.tsx
@@ -8,16 +8,11 @@ interface Props {
 // This component should have an edit button someday.
 export default function ({ username }: Props) {
     return (
-        <div className="bg-black rounded-3 bg-opacity-75 p-3 library-section-container mb-2 pt-4">
+        <div className="basic-container p-3 w-100 mb-2 pt-4">
             <div className="d-flex flex-wrap mb-2">
                 <div className="d-flex justify-content-start flex-grow-1">
-                    <div className="fw-bold">
-                        Biblioteca de <br />
-                        <span className="text-secondary lead">{username}</span>
-                    </div>
+                    <div className="fw-bold">Sua biblioteca</div>
                 </div>
-
-                <div className="d-flex justify-content-end me-3"></div>
             </div>
         </div>
     );