+++ src/components/library/LibrarySection.tsx
@@ -3,6 +3,7 @@ import LibraryBookFigure from "./LibraryBookFigure";
 import React from "react";
 import { Link } from "react-router-dom";
 import { Book } from "../general/helpers/generalTypes";
+import SmallLine from "../general/SmallLine";
 
 interface Props {
     expanded?: boolean;
@@ -22,8 +23,8 @@ export default function LibrarySection(props: Props) {
     };
     const booksInfo = bookCategorySetter();
     return (
-        <div className="d-flex flex-wrap justify-content-center mb-4">
-            <div className="bg-black p-2 rounded-3 bg-opacity-75 text-light p-3 library-section-container">
+        <div className="d-flex flex-wrap justify-content-center mb-4 w-100">
+            <div className="bg-black p-2 rounded-3 bg-opacity-75 text-light p-3 w-100">
                 <div className="d-flex flex-wrap justify-content-center mb-2">
                     <span className="fw-bold lead">{props.message}</span>
                 </div>
@@ -31,68 +32,27 @@ export default function LibrarySection(props: Props) {
                 <div className="d-flex flex-wrap justify-content-center">
                     {booksInfo.length > 0 ? (
                         booksInfo.map((el, i) => {
-                            if (!props.expanded) {
-                                if (i < 4) {
-                                    let timeout;
-                                    i === 0
-                                        ? (timeout = 1000)
-                                        : (timeout = i * 1500);
-                                    return (
-                                        <div
-                                            key={i * Math.random() * 100}
-                                            id="figure-div"
-                                            className="w-100"
-                                        >
-                                            <LibraryBookFigure
-                                                book={el}
-                                                timeout={timeout}
-                                                bookCategory={
-                                                    props.bookCategory
-                                                }
-                                                setProgress={props.setProgress}
-                                            />
-                                        </div>
-                                    );
-                                }
-                            } else {
-                                let timeout;
-                                i === 0
-                                    ? (timeout = 1000)
-                                    : (timeout = i * 1500);
-                                return (
-                                    <div
-                                        key={i * Math.random() * 100}
-                                        id="figure-div"
-                                        className="w-100"
-                                    >
-                                        <LibraryBookFigure
-                                            book={el}
-                                            timeout={timeout}
-                                            bookCategory={props.bookCategory}
-                                            setProgress={props.setProgress}
-                                        />
-                                    </div>
-                                );
-                            }
+                            let timeout;
+                            i === 0 ? (timeout = 1000) : (timeout = i * 1500);
+                            return (
+                                <LibraryBookFigure
+                                    key={i * Math.random() * 100}
+                                    book={el}
+                                    timeout={timeout}
+                                    bookCategory={props.bookCategory}
+                                    setProgress={props.setProgress}
+                                />
+                            );
                         })
                     ) : (
                         <>
-                            <div className="border-white border-top flex-grow-1 p-1" />
+                            <SmallLine flexGrow />
                             <Break />
                             <span className="p-2">
                                 Vazio, que tal adicionar alguns livros?
                             </span>
                         </>
                     )}
-                    {!props.expanded ? (
-                        booksInfo.length > 5 ? (
-                            <div className="d-flex justify-content-start">
-                                <Link to={props.bookCategory}>
-                                    E mais {booksInfo.length - 5}
-                                </Link>
-                            </div>
-                        ) : null
-                    ) : null}
                 </div>
             </div>
         </div>